+++ backend-shared/plannershared.cpp
